// src/utilities/types/operators.ts
export enum Operator {
    Prior = 'PRIOR',
    Is = 'IS',
    Not = '!',
    And = '&&',
    Or = '||',
    Plus = "+",
    Minus = "-",
    MultiplyBy = "*",
    DividedBy = "/",
    EqualTo = '==',
    NotEqualTo = '!=',
    GreaterThan = '>',
    NoLessThan = '>=',
    LessThan = '<',
    NoGreaterThan = '<='
}